우리식당 정상영업합니다 - 컴포넌트 자동 직렬화 역직렬화 에디터 띄우기 까지 됨 드래그안됨

+++ DTEngine/Material.cpp
@@ -294,16 +294,15 @@ Vector4 Material::GetColor() const
     return Vector4(m_data.Color);
 }
 
-int Material::GetShaderID()
+uint16_t Material::GetShaderID()
 {
     return m_shader->GetID();
 }
 
 //Texture는 여러 장이니깐. 조합한 ID를 반환.
-int Material::GetTextureID()
+uint64_t Material::GetTextureID()
 {
     return m_textureBatchID;
-
 }
 
 void Material::UpdateTextureBatchID()
